Update to student CSS
Hey, So I broke the rules and  I merged this. I had to make a change to it and I didn't think I could edit it in its proposed non-merge state? Anyways, this fixes the <input> windows being hidden in the background of the chalkBoard. It was like, bothering my all night then this morning I had a moment when I remembered the classic z-index trick.